requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1     
certifi==2022.12.7
charset-normalizer==2.1.1  
click==8.1.3
colorama==0.4.6
Django==4.1.4
djangorestframework==3.14.0
Flask==2.2.2
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0  
itsdangerous==2.1.2        
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.7
pywhatkit==5.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
zipp==3.11.0